requirements.txt
alembic==1.16.5
altgraph==0.17.4
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-heroku==0.3.1
django-select2==8.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dnspython==2.7.0
email-validator==2.1.0
et_xmlfile==2.0.0
exchangelib==5.5.1
Faker==38.2.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
freetype-py==2.5.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
selenium==4.27.1
webdriver-manager==4.0.2
holidays==0.75
html5lib==1.1
httplib2==0.22.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.7
MouseInfo==0.1.3
msal==1.33.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.1
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==11.2.1
playwright==1.54.0
proto-plus==1.26.1
protobuf==6.31.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycairo==1.29.0
pycparser==2.22
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.26.5
pyparsing==3.2.3
pypdf==6.1.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyspnego==0.11.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32-ctypes==0.2.3
pyxlsb==1.0.10
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rlPyCairo==0.4.0
rsa==4.9.1
## removida duplicata de selenium (mantida a versão 4.27.1 acima)
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.1
sspilib==0.3.1
svglib==1.6.0
tinycss2==1.5.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
watchdog==6.0.0
## removida duplicata de webdriver-manager (mantida a versão 4.0.2 acima)
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.7.0
wikipedia==1.4.0
wsproto==1.2.0
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
